+++ CotEditor/Sources/EditorTextView.swift
@@ -46,6 +46,16 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
     static let didLiveChangeSelectionNotification = Notification.Name("TextViewDidLiveChangeSelectionNotification")
     
     
+    // MARK: Structs
+    
+    private struct SerializationKey {
+        
+        static let insertionLocations = "insertionLocations"
+        
+        private init() { }
+    }
+    
+    
     // MARK: Public Properties
     
     var isAutomaticTabExpansionEnabled = false
@@ -74,6 +84,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
     // MARK: Private Properties
     
     private let matchingBracketPairs: [BracePair] = BracePair.braces + [.doubleQuotes]
+    private lazy var braceHighlightTask = Debouncer(delay: .seconds(0)) { [unowned self] in self.highlightMatchingBrace() }  // NSTextView cannot be weak
     
     private var cursorType: CursorType = .bar
     private var balancesBrackets = false
@@ -192,6 +203,26 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
             #keyPath(scale),
             #keyPath(tabWidth),
         ]
+    }
+    
+    
+    /// store UI state
+    override func encodeRestorableState(with coder: NSCoder) {
+        
+        super.encodeRestorableState(with: coder)
+        
+        coder.encode(self.insertionLocations, forKey: SerializationKey.insertionLocations)
+    }
+    
+    
+    /// restore UI state
+    override func restoreState(with coder: NSCoder) {
+        
+        super.restoreState(with: coder)
+        
+        if let insertionLocations = coder.decodeObject(forKey: SerializationKey.insertionLocations) as? [Int] {
+            self.insertionLocations = insertionLocations
+        }
     }
     
     
@@ -521,9 +552,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
             else { return super.insertNewline(sender) }
         
         let tab = self.isAutomaticTabExpansionEnabled ? String(repeating: " ", count: self.tabWidth) : "\t"
-        
         let ranges = self.rangesForUserTextChange as? [NSRange] ?? [self.rangeForUserTextChange]
-        self.setSelectedRangesWithUndo(ranges)
         
         let indents: [(range: NSRange, indent: String, insertion: Int)] = ranges
             .map { range in
@@ -540,7 +569,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
                 // smart indent
                 if self.isSmartIndentEnabled {
                     let lastCharacter = self.character(before: range)
-                    let nextCharacter = self.character(after: self.rangeForUserTextChange)
+                    let nextCharacter = self.character(after: range)
                     let indentBase = indent
                     
                     // increase indent level
@@ -679,8 +708,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
         if !stillSelectingFlag, !self.isShowingCompletion {
             // highlight matching brace
             if UserDefaults.standard[.highlightBraces] {
-                let bracePairs = BracePair.braces + (UserDefaults.standard[.highlightLtGt] ? [.ltgt] : [])
-                self.highligtMatchingBrace(candidates: bracePairs)
+                self.braceHighlightTask.schedule()
             }
             
             // invalidate current instances highlight
@@ -1518,6 +1546,15 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
             else { return layoutManager.spaceWidth }
         
         return layoutManager.boundingRect(forGlyphRange: NSRange(glyphIndex...glyphIndex), in: textContainer).width
+    }
+    
+    
+    /// highlight the brace matching to the brace next to the cursor
+    private func highlightMatchingBrace() {
+        
+        let bracePairs = BracePair.braces + (UserDefaults.standard[.highlightLtGt] ? [.ltgt] : [])
+        
+        self.highligtMatchingBrace(candidates: bracePairs)
     }