+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Change Log
 - Fix an issue on macOS 10.14 where spaces at the end of the document could not be deleted by delete key.
 - Fix an issue on macOS 10.14 where performing return key just after an open bracket at the end of the document made the application freeze.
 - Fix an issue where the teditor theme for newly opened windows did not match to the window apperance when the system appearance was changed after the application launch.
+- Fix an issue where a highlighting indicator showed up at a wrong location when pressing the enter key just before a closing bracket.
 - [trivial] Fix French localization.