+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 Change Log
 ==========================
 
+3.8.3 (379)
+--------------------------
+
+### Improvements
+
+- Restore all of the last cursors on the window restoration.
+- Highlight matching braces for all cursors.
+- Adjust the printing area by the vertical text orientation.
+- Update JSON syntax style by adding “.resolved” extension.
+
+
+### Fixes
+
+- Fix area to draw text on printing, especially for macOS 10.15 Catalina.
+- Fix an issue on macOS 10.14 where spaces at the end of the document could not be deleted by delete key.
+- Fix an issue on macOS 10.14 where performing return key just after an open bracket at the end of the document made the application freeze.
+- Fix an issue where the editor theme for newly opened windows did not match to the window appearance when the system appearance was changed after the application launch.
+- Fix an issue where a highlighting indicator showed up at a wrong location when pressing the enter key just before a closing bracket.
+- [trivial] Fix French localization.
+
+
+
 3.8.2 (377)
 --------------------------
 
@@ -94,7 +116,7 @@ no change.
 ### Improvements
 
 - Keep last opacity state of restored document windows.
-- Update behavior of Opt+Arrow shortcut series. 
+- Update behavior of Opt+Arrow shortcut series.
 
 
 ### Fixes